Form input control separated into two different helper files - one responsible for user login, one for error display
